Allow extensions to load native libraries
Currently the classloader that gets created when an extension is loaded doesn't pass a path for the extension's native libraries, which means System.loadLibrary will only be able to search for libraries in system directories and not for libraries packaged in the extension. This change would pass through the extension APK's native libraries directory, allowing extensions that include native libraries to load those native libraries.

 it's already technically possible for extensions to load native libraries by creating their own classloader and using reflection to create instances of classes provided by the extension under that classloader. I did that to work around this limitation when experimenting with running AI models in an extension (which require native libraries for performance). Including the extension APK's native library path in the extension classloader just makes it so that extension developers don't need to jump through those hoops.